Migrate location of test of _currency_column_to_numeric
